+++ codec-http2/src/main/java/io/netty/handler/codec/http2/draft10/frame/decoder/Http2FrameDecoder.java
@@ -15,12 +15,6 @@
 
 package io.netty.handler.codec.http2.draft10.frame.decoder;
 
-import static io.netty.handler.codec.http2.draft10.Http2Error.PROTOCOL_ERROR;
-import static io.netty.handler.codec.http2.draft10.Http2Exception.format;
-import static io.netty.handler.codec.http2.draft10.frame.Http2FrameCodecUtil.FRAME_HEADER_LENGTH;
-import static io.netty.handler.codec.http2.draft10.frame.Http2FrameCodecUtil.FRAME_LENGTH_MASK;
-import static io.netty.handler.codec.http2.draft10.frame.Http2FrameCodecUtil.connectionPrefaceBuf;
-import static io.netty.handler.codec.http2.draft10.frame.Http2FrameCodecUtil.readUnsignedInt;
 import io.netty.buffer.ByteBuf;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.ByteToMessageDecoder;
@@ -30,6 +24,8 @@ import io.netty.handler.codec.http2.draft10.frame.Http2Frame;
 import io.netty.handler.codec.http2.draft10.frame.Http2FrameHeader;
 
 import java.util.List;
+
+import static io.netty.handler.codec.http2.draft10.frame.Http2FrameCodecUtil.*;
 
 /**
  * Decodes {@link Http2Frame} objects from an input {@link ByteBuf}. The frames that this handler
@@ -42,14 +38,12 @@ import java.util.List;
 public class Http2FrameDecoder extends ByteToMessageDecoder {
 
     private enum State {
-        PREFACE,
         FRAME_HEADER,
         FRAME_PAYLOAD,
         ERROR
     }
 
     private final Http2FrameUnmarshaller frameUnmarshaller;
-    private final ByteBuf preface;
     private State state;
     private int payloadLength;
 
@@ -62,23 +56,13 @@ public class Http2FrameDecoder extends ByteToMessageDecoder {
             throw new NullPointerException("frameUnmarshaller");
         }
         this.frameUnmarshaller = frameUnmarshaller;
-        preface = connectionPrefaceBuf();
-        state = State.PREFACE;
+        state = State.FRAME_HEADER;
     }
 
     @Override
     protected void decode(ChannelHandlerContext ctx, ByteBuf in, List<Object> out) throws Exception {
         try {
             switch (state) {
-            case PREFACE:
-                processHttp2Preface(ctx, in);
-                if (state == State.PREFACE) {
-                    // Still processing the preface.
-                    break;
-                }
-
-                // Successfully processed the HTTP2 preface.
-
             case FRAME_HEADER:
                 processFrameHeader(in);
                 if (state == State.FRAME_HEADER) {
@@ -102,30 +86,6 @@ public class Http2FrameDecoder extends ByteToMessageDecoder {
             ctx.fireExceptionCaught(t);
             state = State.ERROR;
         }
-    }
-
-    private void processHttp2Preface(ChannelHandlerContext ctx, ByteBuf in) throws Http2Exception {
-        int prefaceRemaining = preface.readableBytes();
-        int bytesRead = Math.min(in.readableBytes(), prefaceRemaining);
-
-        // Read the portion of the input up to the length of the preface, if reached.
-        ByteBuf sourceSlice = in.readSlice(bytesRead);
-
-        // Read the same number of bytes from the preface buffer.
-        ByteBuf prefaceSlice = preface.readSlice(bytesRead);
-
-        // If the input so far doesn't match the preface, break the connection.
-        if (bytesRead == 0 || !prefaceSlice.equals(sourceSlice)) {
-            throw format(PROTOCOL_ERROR, "Invalid HTTP2 preface");
-        }
-
-        if ((prefaceRemaining - bytesRead) > 0) {
-            // Wait until the entire preface has arrived.
-            return;
-        }
-
-        // Start processing the first header.
-        state = State.FRAME_HEADER;
     }
 
     private void processFrameHeader(ByteBuf in) throws Http2Exception {
