+++ codec-http2/src/main/java/io/netty/handler/codec/http2/draft10/frame/encoder/Http2FrameEncoder.java
@@ -15,10 +15,7 @@
 
 package io.netty.handler.codec.http2.draft10.frame.encoder;
 
-import static io.netty.handler.codec.http2.draft10.frame.Http2FrameCodecUtil.connectionPrefaceBuf;
 import io.netty.buffer.ByteBuf;
-import io.netty.channel.ChannelFuture;
-import io.netty.channel.ChannelFutureListener;
 import io.netty.channel.ChannelHandlerContext;
 import io.netty.handler.codec.MessageToByteEncoder;
 import io.netty.handler.codec.http2.draft10.frame.Http2Frame;
@@ -33,7 +30,6 @@ import io.netty.handler.codec.http2.draft10.frame.Http2Frame;
 public class Http2FrameEncoder extends MessageToByteEncoder<Http2Frame> {
 
     private final Http2FrameMarshaller frameMarshaller;
-    private boolean prefaceWritten;
 
     public Http2FrameEncoder() {
         this(new Http2StandardFrameMarshaller());
@@ -47,46 +43,12 @@ public class Http2FrameEncoder extends MessageToByteEncoder<Http2Frame> {
     }
 
     @Override
-    public void channelActive(ChannelHandlerContext ctx) throws Exception {
-        // The channel just became active - send the HTTP2 connection preface to the remote
-        // endpoint.
-        sendPreface(ctx);
-
-        super.channelActive(ctx);
-    }
-
-    @Override
-    public void handlerAdded(ChannelHandlerContext ctx) throws Exception {
-        // This handler was just added to the context. In case it was handled after
-        // the connection became active, send the HTTP2 connection preface now.
-        sendPreface(ctx);
-    }
-
-    @Override
     protected void encode(ChannelHandlerContext ctx, Http2Frame frame, ByteBuf out)
             throws Exception {
         try {
             frameMarshaller.marshall(frame, out, ctx.alloc());
         } catch (Throwable t) {
             ctx.fireExceptionCaught(t);
-        }
-    }
-
-    /**
-     * Sends the HTTP2 connection preface to the remote endpoint, if not already sent.
-     */
-    private void sendPreface(final ChannelHandlerContext ctx) {
-        if (!prefaceWritten && ctx.channel().isActive()) {
-            prefaceWritten = true;
-            ctx.writeAndFlush(connectionPrefaceBuf()).addListener(new ChannelFutureListener() {
-                @Override
-                public void operationComplete(ChannelFuture future) throws Exception {
-                    if (!future.isSuccess() && ctx.channel().isOpen()) {
-                        // The write failed, close the connection.
-                        ctx.close();
-                    }
-                }
-            });
         }
     }
 }
