+++ example/src/main/java/io/netty/example/http2/client/Http2ClientInitializer.java
@@ -15,7 +15,6 @@
  */
 package io.netty.example.http2.client;
 
-import static io.netty.util.internal.logging.InternalLogLevel.INFO;
 import io.netty.channel.ChannelInitializer;
 import io.netty.channel.ChannelPipeline;
 import io.netty.channel.SimpleChannelInboundHandler;
@@ -24,10 +23,13 @@ import io.netty.example.securechat.SecureChatSslContextFactory;
 import io.netty.handler.codec.http2.draft10.connection.Http2ConnectionHandler;
 import io.netty.handler.codec.http2.draft10.frame.Http2DataFrame;
 import io.netty.handler.codec.http2.draft10.frame.Http2FrameCodec;
+import io.netty.handler.codec.http2.draft10.frame.encoder.Http2ClientPrefaceWriter;
 import io.netty.handler.ssl.SslHandler;
 import org.eclipse.jetty.npn.NextProtoNego;
 
 import javax.net.ssl.SSLEngine;
+
+import static io.netty.util.internal.logging.InternalLogLevel.*;
 
 /**
  * Configures the client pipeline to support HTTP/2 frames.
@@ -50,6 +52,7 @@ public class Http2ClientInitializer extends ChannelInitializer<SocketChannel> {
         ChannelPipeline pipeline = ch.pipeline();
 
         pipeline.addLast("ssl", new SslHandler(engine));
+        pipeline.addLast("http2ClientPrefaceWriter", new Http2ClientPrefaceWriter());
         pipeline.addLast("http2FrameCodec", new Http2FrameCodec());
         pipeline.addLast("http2FrameLogger", new Http2FrameLogger(INFO));
         pipeline.addLast("http2ConnectionHandler", new Http2ConnectionHandler(false));
